[script][west conf] Updating west configuration and update_nxp_sdk script to reduce the size of the clone nxp sdk example repo
The aim of this changes is to remove the SDK example repo initialization from the west initialization and instead to use the get sparse-checkout mechanism to reduce the size of the nxp sdk example repo.
The new cloned nxp sdk example repo will contains only required nxp sdk example file that Matter is using others won't be saved locally.

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>