[mcxw7x][platform][examples] Fix Matter CLI not available in some Matter apps
Matter CLI is based on the NXP SDK SerialManager and this component
was not properly initialized in MCXW71 Matter apps due to some missing
build flags. This update fixes the issue and sets a common rule in
using the serial ports by MCXW7x platform Matter apps.

Whenever enabled, Matter CLI will ALWAYS be available on the default
UART1 port. The device logs will be available either on the default
UART1 serial port if Matter CLI is not enabled, or on the auxiliary
UART0 serial port if UART1 is already occupied by the Matter CLI.

This rule aims to provide the most functionality of the Matter apps on
the MCXW7x platforms without needing extra hardware to access auxiliary
consoles.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>