[windows] Updating update_nxp_sdk.py script to clone NXP SDK at the root of matter repo
Goal of this change is to clone the sdk repo to the root of the matter repo in order to reduce SDK path size.
In fact some issue may happen when building Matter on windows if a gcc command line is too long.
To reduce path len SDK is now clone at the matter root on windows.

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>