[cmake][mcxw7x] Adjust build files to build "MbedTLS" using CMake
In the context where the Matter apps are built using 2 different
build systems - a CMake build controlling a "gn" build - up until this
point the "MbedTLS" component was compiled using "gn" and included in
the Matter platform support.

This upgrade restructures the build so that the "MbedTLS" is compiled
by the outer CMake build process instead for enhanced flexibility.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
Reviewed-by: Dina Benamar El Maaroufi <dina.benamarelmaaroufi@nxp.com>